+++ R/plot_comp_bar.R
@@ -4,7 +4,7 @@
 #' - sample_order: Either specify a list of sample names to order manually, or the bars/samples can/will be sorted by similarity, according to a specified distance measure (default aitchison),
 #' - seriate_method specifies a seriation/ordering algorithm (default Ward hierarchical clustering with optimal leaf ordering, see seriation::list_seriation_methods())
 #' - group_by: You can group the samples on distinct plots by levels of a variable in the phyloseq object. The list of ggplots produced can be arranged flexibly with the patchwork package functions. If you want to group by several variables you can create an interaction variable with interaction(var1, var2) in the phyloseq sample_data BEFORE using plot_comp_bar.
-#' - facet_by can allow faceting of your plot(s) by a grouping variable. Using this approach is less flexible than using group_by but means you don't have to arrange a list of plots yourself like with the group_by argument. Using facet_by is equivalent to adding a call to facet_wrap(facets = facet_by, scales = "free") to your plot(s). Calling facet_wrap() yourself is itself a more flexible option as you can add other arguments like the number of rows etc.
+#' - facet_by can allow faceting of your plot(s) by a grouping variable. Using this approach is less flexible than using group_by but means you don't have to arrange a list of plots yourself like with the group_by argument. Using facet_by is equivalent to adding a call to facet_wrap(facets = facet_by, scales = "free") to your plot(s). Calling facet_wrap() yourself is itself a more flexible option as you can add other arguments like the number of rows etc. However you must use keep_all_vars = TRUE if you will add faceting manually.
 #' - bar_width: No gaps between bars, unless you want them (decrease width argument to add gaps between bars).
 #' - bar_outline_colour: Bar outlines default to "black". Set to NA if you don't want outlines.
 #' - palette: Default colouring is consistent across multiple plots if created with the group_by argument, and the defaults scheme retains the colouring of the most abundant taxa irrespective of n_taxa
@@ -15,7 +15,7 @@
 #' @param tax_order order of taxa within the bars, currently only "abundance" works, which puts the most abundant taxa at the bottom (or left).
 #' @param taxon_renamer function that takes taxon names and returns modified names for legend
 #' @param palette palette for taxa fill colours
-#' @param sample_order vector of sample names, any distance measure in calc_dist that does not require a phylogenetic tree, or "default" for the order returned by phyloseq::sample_names(ps)
+#' @param sample_order vector of sample names, any distance measure in dist_calc that does not require a phylogenetic tree, or "default" for the order returned by phyloseq::sample_names(ps)
 #' @param order_with_all_taxa if TRUE, this will use all taxa (not just the top n_taxa) to calculate distances for sample ordering
 #' @param tax_transform_for_ordering transformation of taxa values used before ordering samples by similarity
 #' @param label could also consider arbitrary annotation with extra info, like in complex heatmap
@@ -23,7 +23,7 @@
 #' @param facet_by facets plots by this variable (must be categorical) - if group_by is also set, the faceting with occur separately in the plot for each group.
 #' @param bar_width default 1 avoids random gapping otherwise seen with many samples (set to something less than 1 to introduce gaps between fewer samples)
 #' @param bar_outline_colour line colour separating taxa and samples (use NA for none)
-#' @param drop_unused_vars speeds up ps_melt but might limit future plot customisation options
+#' @param keep_all_vars slows down ps_melt but is required for future plot customisation options like faceting
 #' @param seriate_method name of any ordering method suitable for distance matrices (see ?seriation::seriate)
 #'
 #' @return ggplot or list of harmonised ggplots
@@ -122,7 +122,7 @@ plot_comp_bar <- function(
                           facet_by = NA,
                           bar_width = 1,
                           bar_outline_colour = "black",
-                          drop_unused_vars = TRUE,
+                          keep_all_vars = TRUE,
                           seriate_method = "OLO_ward") {
 
   # check phyloseq for common problems (and fix or message about this)
@@ -168,8 +168,8 @@ plot_comp_bar <- function(
     })
   }
 
-  # drop unused variables
-  if (identical(drop_unused_vars, TRUE)) {
+  # drop unused variables if requested
+  if (identical(keep_all_vars, FALSE)) {
     kept_vars <- label
     if (!identical(group_by, NA)) {
       kept_vars <- c(union(kept_vars, group_by))