+++ R/ps_seriate.R
@@ -1,10 +1,10 @@
 #' Arrange samples in a phyloseq by microbiome similarity
 #'
-#' Uses seriation methods from seriation::seriate and often calc_dist (depending on if seriation method requires a distance matrix)
+#' Uses seriation methods from seriation::seriate and often dist_calc (depending on if seriation method requires a distance matrix)
 #'
 #' @param ps phyloseq object
 #' @param method seriation method for ordering samples, from seriation::seriate
-#' @param dist distance method for calc_dist (only used if required for particular seriation method!)
+#' @param dist distance method for dist_calc (only used if required for particular seriation method!)
 #' @param tax_transform transformation to apply before seriation or any distance calculation
 #' @param add_variable add a variable to the sample data indicating seriation order
 #'
@@ -37,7 +37,7 @@ ps_seriate <- function(ps, method = "OLO_ward", dist = "bray", tax_transform = "
       ser <- seriation::seriate(x = otu_mat, method = method)
     } else if (method %in% seriation::list_seriation_methods(kind = "dist")) {
       # calculate distance between samples
-      distMat <- calc_dist(data = ps_transformed, dist = dist)[["distMat"]]
+      distMat <- dist_calc(data = ps_transformed, dist = dist)[["distMat"]]
       ser <- seriation::seriate(x = distMat, method = method)
     }
     s_order <- seriation::get_order(ser)