+++ R/tax_model_corncob.R
@@ -4,7 +4,7 @@
 #' Same independent variables for all models. Same variables for modelling both the abundance and dispersion parameters.
 #' P-values are from wald tests for simplicity and speed... maybe I will change this in the future.
 #'
-#' @details `model_tax_corncob` can use parallel processing with the `future` package.
+#' @details `tax_model_corncob` can use parallel processing with the `future` package.
 #' This speeds up analysis if you have many taxa to model.
 #' Run a line like this beforehand: `future::plan(future::multisession, workers = 3)`
 #'
@@ -46,7 +46,7 @@
 #' # specify variables used for modelling
 #' variables <- c("female", "overweight", "obese")
 #'
-#' # model_tax_corncob can use parallel processing with the futures package
+#' # tax_model_corncob can use parallel processing with the futures package
 #' # this speeds up analysis if you have many taxa to model
 #' # e.g. # future::plan(future::multisession, workers = 3)
 #'
@@ -56,7 +56,7 @@
 #'   X = phyloseq::rank_names(ps)[-1],
 #'   function(r) {
 #'     message(Sys.time(), " - modelling at level: ", r)
-#'     models <- model_tax_corncob(ps, tax_level = r, variables = variables, taxa = "all")
+#'     models <- tax_model_corncob(ps, tax_level = r, variables = variables, taxa = "all")
 #'     return(models)
 #'   }
 #' )
@@ -67,8 +67,8 @@
 #'
 #' # future::plan(future::sequential) # to turn parallel processing back off
 #' @export
-#' @rdname model_tax_corncob
-model_tax_corncob <- function(
+#' @rdname tax_model_corncob
+tax_model_corncob <- function(
                               ps,
                               tax_level,
                               variables,
@@ -120,13 +120,13 @@ model_tax_corncob <- function(
 
 #' @title Extract stats from corncob taxon model list
 #'
-#' @description Use this function with the output of `model_tax_corncob`. `models2stats_corncob` splits the statistical results extracted from corncob models and groups them by the independent variable name that they refer to. One dataframe of this list can then be joined to the output of taxatree_nodes to prepare for taxonomic heat tree graph visualisation of taxon-variable associations.
+#' @description Use this function with the output of `tax_model_corncob`. `models2stats_corncob` splits the statistical results extracted from corncob models and groups them by the independent variable name that they refer to. One dataframe of this list can then be joined to the output of taxatree_nodes to prepare for taxonomic heat tree graph visualisation of taxon-variable associations.
 #'
-#' @param taxon_models named list output of `model_tax_corncob`
+#' @param taxon_models named list output of `tax_model_corncob`
 #'
 #' @return list of dataframes, one df per independent variable
 #' @export
-#' @rdname model_tax_corncob
+#' @rdname tax_model_corncob
 models2stats_corncob <- function(taxon_models) {
   # get stats from models
   taxon_stats <- lapply(taxon_models, corncob::waldt)